+++ examples/javascript/02_googleassistantconv/typeoverride/src/config.js
@@ -1,0 +1,13 @@
+const { config } = require('jovo-framework');
+
+module.exports = config({
+	logging: true,
+	intentMap: {
+		'AMAZON.HelpIntent': 'HelpIntent'
+	},
+	db: {
+		FileDb: {
+			pathToFile: './../db/db.json'
+		}
+	},
+});