+++ jovo-core/src/core/Jovo.ts
@@ -10,6 +10,7 @@ import {
   ErrorCode,
   JovoError,
   Log,
+  QuickReply,
   SessionConstants,
   SessionData,
   SpeechBuilder,
@@ -360,9 +361,7 @@ export abstract class Jovo extends EventEmitter {
     }
 
     delete this.$output.ask;
-    this.$output.tell = {
-      speech: speech.toString(),
-    };
+    _set(this.$output, 'tell.speech', speech.toString());
     return this;
   }
 
@@ -393,11 +392,14 @@ export abstract class Jovo extends EventEmitter {
       reprompt = speech;
     }
 
-    this.$output.ask = {
-      speech: speech.toString(),
-      reprompt: reprompt.toString(), // tslint:disable-line:object-literal-sort-keys
-    };
+    _set(this.$output, 'ask.speech', speech.toString());
+    _set(this.$output, 'ask.reprompt', reprompt.toString());
 
+    return this;
+  }
+
+  showQuickReplies(quickReplies: Array<QuickReply | string>) {
+    _set(this.$output, 'quickReplies', quickReplies);
     return this;
   }