+++ jovo-core/src/index.ts
@@ -1,7 +1,6 @@
 import { AppConfig } from './Interfaces';
 import { Log } from './util/Log';
 import { Project as P } from './util/Project';
-
 export * from './plugins/types';
 
 export const Project = P.getInstance();
@@ -70,6 +69,7 @@ export {
   ComponentResponseStatus,
 } from './plugins/Component';
 export { Router } from './plugins/Router';
+
 export * from './Interfaces';
 
 export * from './util/HttpService';