+++ jovo-integrations/jovo-db-postgresql/src/index.ts
@@ -1,0 +1,12 @@
+export { PostgreSQL, Config as PostgreSQLConfig } from './PostgreSQL';
+
+import { Config } from './PostgreSQL';
+
+interface AppPostgreSQLConfig {
+  PostgreSQL?: Config;
+}
+
+declare module 'jovo-core/dist/src/Interfaces' {
+  export interface AppDbConfig extends AppPostgreSQLConfig {}
+  export interface ExtensiblePluginConfigs extends AppPostgreSQLConfig {}
+}