+++ jovo-platforms/jovo-platform-googleassistantconv/src/GoogleAssistant.ts
@@ -18,13 +18,10 @@ import { ConversationalActionsCore } from './modules/ConversationalActionsCore';
 import { GoogleAssistantTestSuite } from './core/Interfaces';
 import { GoogleAction } from './core/GoogleAction';
 import { MediaResponsePlugin } from './modules/MediaResponse';
+import { InteractiveCanvas } from './modules/InteractiveCanvas';
 
 export interface Config extends ExtensibleConfig {
   handlers?: any; //tslint:disable-line
-  transactions?: {
-    androidPackageName?: string;
-    keyFile?: object;
-  };
 }
 
 export class GoogleAssistant extends Platform<
@@ -33,7 +30,6 @@ export class GoogleAssistant extends Platform<
 > {
   config: Config = {
     enabled: true,
-    conversationalActions: false,
     plugin: {},
   };
 
@@ -82,11 +78,6 @@ export class GoogleAssistant extends Platform<
       return this.constructor.name === 'GoogleAction';
     };
 
-    /**
-     * Sets alexa handlers
-     * @public
-     * @param {*} handlers
-     */
     // tslint:disable-next-line
     BaseApp.prototype.setGoogleAssistantHandler = function (...handlers: any[]) {
       for (const obj of handlers) {
@@ -99,7 +90,7 @@ export class GoogleAssistant extends Platform<
       }
       return this;
     };
-    this.use(new ConversationalActionsCore(), new MediaResponsePlugin());
+    this.use(new ConversationalActionsCore(), new MediaResponsePlugin(), new InteractiveCanvas());
   }
 
   makeTestSuite(): GoogleAssistantTestSuite {