+++ jovo-clients/jovo-client-web/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Current version might be higher than the latest changes displayed below because of updates of dependencies.
 
+
+### 2020-11-05 [3.2.1]`
+* [#848](https://github.com/jovotech/jovo-framework/pull/848) Improve Web-Client browser-detection ([@m-ripper](https://github.com/m-ripper))
+
+
+
 ## 3.0.18 (2020-05-13)
 
 #### :bug: Bug Fix